added git logo
http://upload.wikimedia.org/wikipedia/commons/e/e0/Git-logo.svg

License:
This work is free software; you can redistribute it and/or modify
it under the terms of the GNU General Public License as published by the
Free Software Foundation; either version 2 of the License, or any later
version. This work is distributed in the hope that it will be useful, but
WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY
or FITNESS FOR A PARTICULAR PURPOSE. See version 2 and version 3 of the
GNU General Public License for more details.