Applied most of Steve's corrections.
Here is the full list...

p28
s/colaboration/collaboration/

p56-59
You have an arrowhead without a line for "time" (from page 55).  Could
be my viewer, but i don't think so.

p82 (starting from p74)
A little cluttered.  Viewers at the back may have to stand up to see
the end of your slide depending on where the screen is.  I suggest
shrinking all the images (on these slides) by about 10%.

p83
You refer to this section as "Decentralization", but the previous
section as "Centralized".  They should match.  Either
"Centralized/Decentralized" or "Centralization/Decentralization".

p100
All the bullet points come up at once.  Not sure if this is what you
want.  If you intend to talk about these points for any length of
time, it seems to violate your rule about having lots of text to read
(in a short period of time) while listening to the speaker.

p105-108
The heading for this is pretty long.  If the projector isn't 100%
calibrated, some of this might run off the side of the screen.

p108
There is space between bullet points.  Probably intentionally, but you
don't use this anywhere else.

p113
s/colaboration/collaboration/

p119-121
Since you have no text, the graphic could be bigger.  The text (in the
graphic) is tiny, and i think more than a few people will try hard to
read it, loosing concentration on your voice.  If not grown, perhaps
just center it.

p142-146 (starting 136)
I get the impression from these slides that the "developers" are
pushing to the server, just in a different directory.  Should the big
contiguous light blue square be multiple squares?

p147-155
These pages were very slow to render.  I'm not sure what your plan is
for presenting, but you should find a way to pre-render them.  My
setup was not able to cache the entire rendered file (i went through
them start to finish twice).

p148-151
The screen color of the lower center laptop changes during these
slides (changes back after).  It implies some difference in state, but
i don't think that's what you meant.

p151
Same as above, with the 3 laptops screens in the lower right.

p153
More of the same as the last two.

p155
And again.

p157-end
The "Intro" section has only one subsection, "Operations".  It
therefore does not need a subsection at all, or the banner in the
upper left marking it as such.

p197
Before you publish this online, you may want to remove the trademarked material.