steve's comments on UG/intro slides
---8<---
p28
The footnote does not appear until p30.  I had issues with this with
beamer too, not sure if you can fix it or want to.  It does make
people think, "What is the * for?", which breaks continuity.  Of
course, perhaps that's what you want.

p34
Two bullet points appear at once.  Based on what you said last time,
this is probably what you want, but just letting you know in case it
isn't.

p62
I think the lettering in the arrow should read top to bottom.  This is
how you want the viewers eye to move.  The way it is now, you have to
go to the bottom, read up, and then go back down.  It caused my brain
to hiccough and i had to concentrate to read it and continue.

p80-94
Is it imperative that the parent entry be in the 2nd place?  The
flow of your talk isn't the same order as the entries (ref p87).  May
be a bit of work to change this though.

p95
Is this page intentionally blank?