steve's comments on UG/cmd slides
---8<---
p97-107
This subsection lacks a title.  May i suggest "Overview", as in
"Commands:Overview"?  Your banner may not fit if you add it the
subsection title though.

p100-106
I like the style of this transition.  Very intuitive.

p124-140
The title page says "Configuration", but the banner says "Config".  I
suspect you are intending for the banner to refer to the config it
"git config".  In that case, it's probably correct as is.

p136
A line runs off the page.  It's just barely off, so it's fine.  Just
keep it in mind if you change the text to be any longer.  For example,
if the name of the alias changes from "st" to "stat" it will overrun
the box.

p141
Is this page intentionally blank?

p143
Is this page intentionally (nearly) blank?