steve's comments on UG/using slides
---8<---
p202-205
The "3fb8a8e" hangs outside the box.  It doesn't do this on p206.

p222
I think the symbolic short SHA1 id here (db35c00^) should be "895740c^".

p263
The title page says "References", but the banner says "Refs".  I can't
come up with a reason for this this time.  I think the banner should
say "References" too, the short form feeling too informal.

p273-284
I like how "git branch" stays in the same place through this entire sequence.

p294-296
I think there is a pacing issue here.  Should only be 2 slides, not 3.

p306-316
Is there a reason that the title (Remotes?) is a question?

p317
p306-316 have the title "Remotes", but this page does not.  Instead it
has the github logo.  I think it belongs in the same subsection, and
should have both the title and the logo.

p318-336
Are these pages intentionally blank?