more comments from Steve
----
"Tracking" slides.

p146
  The "push" and "pull" operations are graphically indistinct.  To me,
it looks like "pull" does the same thing as "push".  You usually use a
different colour arrow for such things.  If you change it, you should
also change the colour for "clone" so it is likewise distinct.

  I read it to mean that "light blue" was pulling from "orange" once
my brain looped on it twice.

  In the immediately previous slides, when you introduce "pull" and
"clone", you have text indicating it is the new information, but not
with this slide.  It's already a little crowded, so i don't know if
consistency is that important to you.

p162
  The Dictator slides are much faster to render, but there is still a
noticeable pause.  Once again, my viewer was not able to cache this
for my second pass through.  Probably nothing can be done at this
point, just wanted to make you aware.

p165
  The two arrows are slightly different shades.  Intentional?

p172
  The "Intro" section has no other subsections than "to Git".  Is it
this way because "Intro to Git" wouldn't fit in the black banner?
Perhaps "Git Intro" (or if you want to be silly "GInTro", maybe with
different colours to make the letter groupings more distinct) for the
section, and no subsection?